% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addCacheSearchEnvironment}
\alias{addCacheSearchEnvironment}
\title{append a new Environment name (a character string) to a global option
which is a vector of such names. SimpleCache will search all of these
environments to check if a cache is previously loaded, before reloading it.}
\usage{
addCacheSearchEnvironment(addEnv)
}
\description{
append a new Environment name (a character string) to a global option
which is a vector of such names. SimpleCache will search all of these
environments to check if a cache is previously loaded, before reloading it.
}

