% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadCaches.R
\name{availCaches}
\alias{availCaches}
\title{Show available caches.}
\usage{
availCaches(cacheSubDir = "")
}
\arguments{
\item{cacheSubDir}{Optional parameter to specify a subdirectory of the cache folder.}
}
\description{
Show available caches.
}

