% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{batchWrap}
\alias{batchWrap}
\title{Wrapper function for batchtools integration}
\usage{
batchWrap(cacheName, instruction, cacheDir, batchDir, batchMethod)
}
\description{
Wrapper function for batchtools integration
}
