% Generated by roxygen2 (4.0.1): do not edit by hand
\name{buildSlurmScript}
\alias{buildSlurmScript}
\title{Given arguments from a slurmParams object, creates and (possibly) submits a
slurm job.}
\usage{
buildSlurmScript(rcode, preamble = "", submit = FALSE,
  hpcFolder = "slurm", jobName = "test", mem = "4000", cores = "1",
  partition = "develop", timeLimit = "02:00:00", sourceProjectInit = TRUE)
}
\description{
Given arguments from a slurmParams object, creates and (possibly) submits a
slurm job.
}

