% Generated by roxygen2 (4.0.1): do not edit by hand
\name{downloadCache}
\alias{downloadCache}
\title{Create or load a cache from the web.}
\usage{
downloadCache(object, url, env = NULL, reload = FALSE, recreate = FALSE,
  noload = FALSE, cacheDir = getOption("RCACHE.DIR"),
  cacheSubDir = "download", loadEnvir = environment())
}
\description{
Given a URL, this function will download the file and cache it, or load it if it has been previously downloaded.
BETA -- not finished.
TODO -- update with searchEnvir feature of simpleCache.
}

