% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheDirectories.R
\name{getCacheDir}
\alias{getCacheDir}
\title{Fetcher of the currently set cache directory.}
\usage{
getCacheDir()
}
\value{
If the option is set, the path to the currently set cache directory; otherwise, \code{NULL}.
}
\description{
\code{getCacheDir} retrieves the value of the option that stores the currently 
set cache directory path.
}
