% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadCaches}
\alias{loadCaches}
\title{This function just takes a list of caches, and loads them. It's designed
for stuff you already cached previously, so it won't build any caches.}
\usage{
loadCaches(cacheNames, ...)
}
\description{
This function just takes a list of caches, and loads them. It's designed
for stuff you already cached previously, so it won't build any caches.
}

