% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resetCacheSearchEnvironment}
\alias{resetCacheSearchEnvironment}
\title{Sets global option of cache search environments to NULL.}
\usage{
resetCacheSearchEnvironment()
}
\description{
Sets global option of cache search environments to NULL.
}

