% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utility.R
\name{secToTime}
\alias{secToTime}
\title{This function takes a time in seconds and converts it to a more
human-readable format, showing hours, minutes, or seconds, depending
on how long the time is. Used by my implementation of tic()/toc().}
\usage{
secToTime(timeInSec)
}
\arguments{
\item{timeInSec}{numeric value of time measured in seconds.}
}
\description{
This function takes a time in seconds and converts it to a more
human-readable format, showing hours, minutes, or seconds, depending
on how long the time is. Used by my implementation of tic()/toc().
}

