% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheDirectories.R
\name{setCacheBuildDir}
\alias{setCacheBuildDir}
\title{Sets local cache build directory with scripts for building files.}
\usage{
setCacheBuildDir(cacheBuildDir)
}
\arguments{
\item{cacheBuildDir}{Directory where build scripts are stored.}
}
\description{
Sets local cache build directory with scripts for building files.
}
