% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheDirectories.R
\name{setSharedCacheDir}
\alias{setSharedCacheDir}
\title{Set shared cache directory}
\usage{
setSharedCacheDir(sharedCacheDir = NULL)
}
\arguments{
\item{sharedCacheDir}{Directory where shared caches should be stored}
}
\description{
Sets global variable specifying the default cache directory for
\code{\link{simpleCacheShared}} calls; this function is simply a helper alias for caching
results that will be used across projects.
}
