% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCache.R
\docType{package}
\name{simpleCache-package}
\alias{simpleCache-package}
\alias{_PACKAGE}
\title{Provides intuitive functions for caching R objects, encouraging faster
reproducible and restartable R analysis}
\description{
Provides intuitive functions for caching R objects, encouraging faster
reproducible and restartable R analysis
}
\references{
\url{https://github.com/databio/simpleCache}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.github.com/databio/simpleCache}
  \item Report bugs at \url{https://www.github.com/databio/simpleCache}
}

}
\author{
Nathan Sheffield
}
