% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simpleCacheGlobal}
\alias{simpleCacheGlobal}
\title{Helper alias for loading caches into the global environment.
simpleCache normally loads variables into the calling environment; this
ensures that the variables are loaded in the global environment.}
\usage{
simpleCacheGlobal(...)
}
\description{
Helper alias for loading caches into the global environment.
simpleCache normally loads variables into the calling environment; this
ensures that the variables are loaded in the global environment.
}

