% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedCaches.R
\name{simpleCacheShared}
\alias{simpleCacheShared}
\title{Alias to default to a shared cache folder.}
\usage{
simpleCacheShared(...)
}
\arguments{
\item{...}{Parameters passed to \code{\link{simpleCache}}.}
}
\description{
Helper alias for caching across experiments/people.
Just sets the cacheDir to the default SHARE directory 
(instead of the typical default PROJECT directory)
}
