% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simpleCacheSharedGlobal}
\alias{simpleCacheSharedGlobal}
\title{Helper alias for loading shared caches into the global environment.}
\usage{
simpleCacheSharedGlobal(...)
}
\description{
Helper alias for loading shared caches into the global environment.
}

