% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCache.R
\name{testExec}
\alias{testExec}
\title{Debugging function... I can probably delete it.}
\usage{
testExec(instruction)
}
\arguments{
\item{instruction}{R code to run.}
}
\description{
Debugging function... I can probably delete it.
}

