% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{tic}
\alias{tic}
\title{Start a timer}
\usage{
tic(gcFirst = TRUE, type = c("elapsed", "user.self", "sys.self"))
}
\arguments{
\item{gcFirst}{Garbage Collect before starting the timer?}

\item{type}{Type of time to return,
can be 'elapsed', 'user.self', or 'sys.self'}
}
\description{
Start a timer
}

