% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheDirectories.R
\name{addCacheSearchEnvironment}
\alias{addCacheSearchEnvironment}
\title{Add a cache search environment}
\usage{
addCacheSearchEnvironment(addEnv)
}
\arguments{
\item{addEnv}{Environment to append to the shared cache search list}
}
\description{
Append a new Environment name (a character string) to a global option
which is a vector of such names. SimpleCache will search all of these
environments to check if a cache is previously loaded, before reloading it.
}
