% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCache.R
\name{downloadCache}
\alias{downloadCache}
\title{Create or load a cache from the web.}
\usage{
downloadCache(object, url, env = NULL, reload = FALSE, recreate = FALSE,
  noload = FALSE, cacheDir = getOption("RCACHE.DIR"),
  cacheSubDir = "download", loadEnvir = parent.frame(),
  assignToVariable = NULL)
}
\arguments{
\item{object}{Name of cache.}

\item{url}{Web location of the text file to cache.}

\item{env}{See documentation at simpleCache()}

\item{reload}{See documentation at simpleCache()}

\item{recreate}{See documentation at simpleCache()}

\item{noload}{See documentation at simpleCache()}

\item{cacheDir}{See documentation at simpleCache()}

\item{cacheSubDir}{See documentation at simpleCache()}

\item{loadEnvir}{See documentation at simpleCache()}

\item{assignToVariable}{See documentation at simpleCache()}
}
\description{
Given a URL, this function will download the file and cache it, or load it if it has been previously downloaded.
BETA -- not finished.
TODO -- update with searchEnvir feature of simpleCache.
}

