% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slurmCache.R
\name{getSlurmParams}
\alias{getSlurmParams}
\title{Returns a slurm parameterization object for use in simpleCache();}
\usage{
getSlurmParams(preamble = "", submit = T, hpcFolder = "slurm",
  jobName = "test", mem = "4000", cores = "5", partition = "develop",
  timeLimit = "02:00:00", sourceProjectInit = TRUE)
}
\description{
Returns a slurm parameterization object for use in simpleCache();
}

