% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadCaches.R
\name{loadCaches}
\alias{loadCaches}
\title{This function just takes a list of caches, and loads them. It's designed
for stuff you already cached previously, so it won't build any caches.}
\usage{
loadCaches(cacheNames, ...)
}
\arguments{
\item{cacheNames}{Vector of caches to load.}

\item{...}{Additional parameters passed to simpleCache.}
}
\description{
This function just takes a list of caches, and loads them. It's designed
for stuff you already cached previously, so it won't build any caches.
}

