% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheDirectories.R
\name{resetCacheSearchEnvironment}
\alias{resetCacheSearchEnvironment}
\title{Sets global option of cache search environments to \code{NULL}.}
\usage{
resetCacheSearchEnvironment()
}
\description{
Sets global option of cache search environments to \code{NULL}.
}
