% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cacheDirectories.R
\name{resetCacheSearchEnvironment}
\alias{resetCacheSearchEnvironment}
\title{Sets global option of cache search environments to NULL.}
\usage{
resetCacheSearchEnvironment()
}
\description{
Sets global option of cache search environments to NULL.
}

