% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cacheDirectories.R
\name{setCacheDir}
\alias{setCacheDir}
\title{Sets a global variable specifying the default cache directory for simpleCache() calls.}
\usage{
setCacheDir(cacheDir)
}
\arguments{
\item{cacheDir}{Directory where caches should be stored}
}
\description{
Sets a global variable specifying the default cache directory for simpleCache() calls.
}

