% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setSharedCacheDir}
\alias{setSharedCacheDir}
\title{Sets global variable specifying the default cache directory for simpleCacheShared() calls; this function is simply a helper alias for caching results that will be used across experiments.}
\usage{
setSharedCacheDir(sharedCacheDir)
}
\description{
Sets global variable specifying the default cache directory for simpleCacheShared() calls; this function is simply a helper alias for caching results that will be used across experiments.
}

