% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cacheDirectories.R
\name{setSharedCacheDir}
\alias{setSharedCacheDir}
\title{Sets global variable specifying the default cache directory for simpleCacheShared() calls; this function is simply a helper alias for caching results that will be used across experiments.}
\usage{
setSharedCacheDir(sharedCacheDir)
}
\arguments{
\item{sharedCacheDir}{Directory where shared caches should be stored}
}
\description{
Sets global variable specifying the default cache directory for simpleCacheShared() calls; this function is simply a helper alias for caching results that will be used across experiments.
}

