% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCache.R
\docType{package}
\name{simpleCache-package}
\alias{simpleCache-package}
\alias{_PACKAGE}
\title{Provides intuitive functions for caching R objects, encouraging faster
reproducible and restartable R analysis}
\description{
Provides intuitive functions for caching R objects, encouraging
    reproducible, restartable, and distributed R analysis. The user selects a
    location to store caches, and then provides  nothing more than a cache name
    and instructions (R code) for how to produce the R object. Also
    provides some advanced options like environment assignments, recreating or
    reloading caches, and cluster compute bindings (using the 'batchtools'
    package) making it flexible enough for use in large-scale data analysis
    projects.
}
\references{
\url{https://github.com/databio/simpleCache}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/databio/simpleCache}
  \item Report bugs at \url{https://github.com/databio/simpleCache}
}

}
\author{
Nathan Sheffield
}
