% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simpleCache.R
\name{simpleCacheGlobal}
\alias{simpleCacheGlobal}
\title{Helper alias for loading caches into the global environment.
simpleCache normally loads variables into the calling environment; this
ensures that the variables are loaded in the global environment.}
\usage{
simpleCacheGlobal(...)
}
\arguments{
\item{...}{Parameters passed to simpleCache().}
}
\description{
Helper alias for loading caches into the global environment.
simpleCache normally loads variables into the calling environment; this
ensures that the variables are loaded in the global environment.
}

