% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simpleCacheShared}
\alias{simpleCacheShared}
\title{Helper alias for caching across experiments/people.
Just sets the cacheDir to the default SHARE directory
(instead of the typical default PROJECT directory)}
\usage{
simpleCacheShared(...)
}
\description{
Helper alias for caching across experiments/people.
Just sets the cacheDir to the default SHARE directory
(instead of the typical default PROJECT directory)
}

