% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCache.R
\name{simpleCacheShared}
\alias{simpleCacheShared}
\title{Helper alias for caching across experiments/people.
Just sets the cacheDir to the default SHARE directory 
(instead of the typical default PROJECT directory)}
\usage{
simpleCacheShared(...)
}
\arguments{
\item{...}{Parameters passed to simpleCache().}
}
\description{
Helper alias for caching across experiments/people.
Just sets the cacheDir to the default SHARE directory 
(instead of the typical default PROJECT directory)
}
