% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simpleCacheSharedGlobal}
\alias{simpleCacheSharedGlobal}
\title{Helper alias for loading shared caches into the global environment.}
\usage{
simpleCacheSharedGlobal(...)
}
\arguments{
\item{...}{Parameters passed to simpleCache().}
}
\description{
Helper alias for loading shared caches into the global environment.
}
